backend/requirements.txt
# Azure AI Foundry SDK
azure-ai-projects>=1.0.0b1
azure-ai-inference>=1.0.0b1
azure-identity>=1.15.0

# Azure Services
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-ai-formrecognizer>=3.3.0
azure-cosmos>=4.5.1

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# LangChain (for agent orchestration)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.2
python-docx>=1.1.0
pypdf>=3.17.0
python-pptx>=0.6.23
pyyaml>=6.0.1
lxml>=5.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1

# Graph (for KAG)
networkx>=3.2.1
gremlinpython>=3.7.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
